Implement deferred wrapping to fix terminal flicker (at very top/bottom, notably with zellij) - also tweak various fns for correctness (mostly as a result of hunting down the flicker bug)
The problem:
When a command like `echo 'text'` finished, and the shell printed a new
prompt, an extra blank line would sometimes be inserted, causing
everything to scroll up one line too many, and also overrite (for
instance) the zellij status bar.

A newline `\n` was being inserted, along with an implicit wrap,
resulting in a double line feed.